fix: useIsmounted build issue
