+++ src/hooks/use-is-mounted.ts
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function useIsMounted() {
+  const [hasMounted, setHasMounted] = React.useState(false);
+
+  React.useEffect(() => {
+    setHasMounted(true);
+  }, []);
+
+  return hasMounted;
+}
